VIRT_UART_Transmit should take const void* pData
The typing of the `pData` argument to `VIRT_UART_Transmit` should match the type of the data argument to `OPENAMP_send` aka `rpmsg_send`, which is `const void*` (not `uint8_t*`). This will allow users to send arbitrary data, including constant data like string literals and data in the form of a struct pointer, without needing to cast the data to a `uint8_t*` first. Additionally, this may allow the compiler to further optimize the `VIRT_UART_Transmit` function.